Proposal: removing default actions
This is more of a discussion than a concrete PR
The plug currently assumes, by default, that only CRUD actions are
worthy of being checked.

This feels like risky behaviour to me.

Someone can easily, and unkowingly use this to authorize their
controller, noticing that it works the regular CRUD actions, but easily
miss that his one custom action is actually not checked at all

I would expect this to check everything by default, and at most,
`default_actions` being an opt-in, rather than an opt-out

In the short term, though, removing that altogether until a use case for
it comes along is a perfectly valid option

Also worth noting: after removing the logic for this, no tests failed at
all. Including the one that was suposed to be checking this exact
behaviour