Removes default policy implementation
Functions tagged with `defoverridable` are not used as a fallback if an
upper level module re-implementes the behaviour. This means that the
following policy

```
defmodule Policy do
  use Dictator.Policy

  def can?(:very, :specific, :scenario), do: false
end
```

Would still fail to run for any other arguments, despite the original
policy module defining a default implementation.

This renders the default implementation pretty much useless, so
I propose we remove it